Rename module from bulk_update to django_bulk_update.
